Do not add taxon if it is already on the product
When processing rows that are variants, this code will get called for
a product that has been previously processed by the master variant row.
In this case, the taxons already exist on the product and be in
the database. This causes a validation error to be raised because the
taxon cannot be added twice to the product. We can get around this by
just checking if the taxon that will be added is already there.