Skip empty option values
The Shopify CSV format specifies the OptionValue column be empty for
the first row of a product with multiple variants. I believe this row
becomes the master variant. I noticed when I was following the Shopify
documentation that the first row for products was failing because it
would attempt to make an option value with the empty string, which
then causes a validation error to be thrown.

Instead we can just skip through any empty values.