Use Paperclip to read the Paperclip attachments
We can't rely on the attachment path to be something we can just
`File.read` because it may not be a local file which can be opened.

Instead we can rely on Paperclip to give us an IO adapter which
is capable of reading an attachment independent from the backend.

Co-Authored-By: Jared Norman <jared@super.gd>
Co-Authored-By: Benjamin Willems <benjamin@super.gd>
Co-Authored-By: Graham Bouvier <graham@super.gd>
Co-Authored-By: Madeline Collier <madeline@super.gd>