Force encoding to UTF-8 when reading paperclip import file
In this call to CSV.parse, we specify that we expect the encoding of the content
we are parsing to be UTF-8.
Under nominal circumstances, this is the case, but if a file is encoded in another
format, the underlying conversion to JSON may attempt to convert the file from the
encoding it says it is to UTF-8.
If the encoding of a file is incorrect, i.e. the file states it is encoded ASCII-8BIT
but the file in fact contains unicode bytes, the conversion will fail and the records
for the import will fail to save.
Since we specify that the encoding will be UTF-8 for the parse, it is sensible to force
reading of the file as UTF-8 regardless of the actual stated encoding of the file.
In almost all cases this will work without issue, as even ASCII can be read as UTF-8
without issue.
Support for non-UTF-8 encodings would take more work to implement and it's not clear
this would be required for any users.

Co-Authored-By: Madeline Collier <madeline@super.gd>
Co-Authored-By: Benjamin Willems <benjamin@super.gd>
Co-Authored-By: Jared Norman <jared@super.gd>
Co-Authored-By: Adam Mueller <adam@super.gd>